feat(darkMode): refactor dark mode handling into a mixin for better code organization
